pyproject.toml
[project]
name = "ml_template"
version = "0.1.0"
requires-python = ">= 3.9"
dependencies = [
  "black==24.1.1",
  "flake8==7.0.0",
  "isort==5.13.2",
  "pre-commit==3.6.0",
  "wandb==0.16.3",
  "hydra-core==1.3.2",
  "matplotlib==3.8.2",
  "scikit-learn==1.4.0",
  "numpy==1.26.4",
  "pytest==8.0.0",
  "torch==2.2.0",
  "lightning==2.2.0",
  "albumentations==1.3.1",
  "segmentation_models_pytorch==0.3.3",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
traintest = "traintest:main"
